added to bash the zip download and unzip, updated test, updated s3hander

+++ Assignment1/src/test/java/Test.java
@@ -42,10 +42,9 @@ public class Test {
         AmazonS3 s3 = S3Handler.connectS3(credentials);
 
 //        testInstances(ec2);
+//        String fileName = "DemoFileToS3.txt";
+//        testS3(credentials, ec2, s3, fileName);
 
-        String fileName = "../../../DemoFileToS3.txt";
-
-        testS3(credentials, ec2, s3, fileName);
 
     }
 
@@ -53,7 +52,7 @@ public class Test {
     public static void testInstances(AmazonEC2 ec2) throws Exception {
         System.out.println("\n\n*** test EC2 ***");
 
-        List<Instance> myInstances = EC2Handler.launchEC2Instances(ec2, 1);
+        List<Instance> myInstances = EC2Handler.launchEC2Instances(ec2, 1, "");
         if(myInstances != null){
             Instance manager = myInstances.get(0);
             String instanceIdToTerminate = manager.getInstanceId();
@@ -75,38 +74,61 @@ public class Test {
     // Test S3 - uploading a file
     public static void testS3(AWSCredentialsProvider credentials, AmazonEC2 ec2, AmazonS3 s3, String fileName) throws Exception {
         System.out.println("\n\n*** test S3 ***");
+        String bucketName = null;
+        String keyName = null;
 
-        System.out.println("\nCreate a bucket");
-        String bucketName = S3Handler.createBucket(credentials, s3, fileName);
+        try {
 
-        System.out.println("\nUpload file to S3");
-        String keyName = S3Handler.uploadFileToS3(ec2, s3, bucketName, fileName);
+            System.out.println("\nCreate a bucket");
+            bucketName = S3Handler.createBucket(credentials, s3, fileName);
 
-        System.out.println("\nListing buckets");
-        for (Bucket bucket : s3.listBuckets()) {
-            System.out.println(" - " + bucket.getName());
-        }
+            System.out.println("\nUpload file to S3");
+            keyName = S3Handler.uploadFileToS3(ec2, s3, bucketName, fileName);
 
-        // This works, but is heavy so I commented it out
+            System.out.println("\nListing buckets");
+            for (Bucket bucket : s3.listBuckets()) {
+                System.out.println(" - " + bucket.getName());
+            }
+
+            // This works, but is heavy so I commented it out
 //        System.out.println("\nDownloading an object");
 //        S3Object object = s3.getObject(new GetObjectRequest(bucketName, keyName));
 //        System.out.println("Content-Type: "  + object.getObjectMetadata().getContentType());
 //        displayTextInputStream(object.getObjectContent());
 
-        // TODO: this doesn't work (unknown reason)
-        System.out.println("\nListing objects");
-        ObjectListing objectListing = s3.listObjects(new ListObjectsRequest()
-                .withBucketName(bucketName)
-                .withPrefix("My"));
-        for (S3ObjectSummary objectSummary : objectListing.getObjectSummaries()) {
-            System.out.println(" - " + objectSummary.getKey() + "  " +
-                    "(size = " + objectSummary.getSize() + ")");
+            System.out.println("\nListing objects");
+            ObjectListing objectListing = s3.listObjects(new ListObjectsRequest()
+                    .withBucketName(bucketName)
+                    .withPrefix(""));
+            for (S3ObjectSummary objectSummary : objectListing.getObjectSummaries()) {
+                System.out.println(" - " + objectSummary.getKey() + "  " +
+                        "(size = " + objectSummary.getSize() + ")");
+            }
         }
+        finally {
 
-        System.out.println("\nDelete file from S3");
-        S3Handler.deleteFile(s3, bucketName, keyName);
+            if (keyName != null) {
+                System.out.println("\nDelete file from S3");
+                S3Handler.deleteFile(s3, bucketName, keyName);
+            }
 
-        System.out.println("\nDelete bucket");
-        S3Handler.deleteBucket(s3, bucketName);
+            if (bucketName != null) {
+                System.out.println("\nDelete bucket");
+                S3Handler.deleteBucket(s3, bucketName);
+            }
+        }
+    }
+
+
+    public static void testJars(AWSCredentialsProvider credentials, AmazonEC2 ec2, AmazonS3 s3) {
+
+        String keyName = "";
+        String bucketName = "akiaj24cwsltdpfv43lqajars";
+
+        System.out.println("Download jars zip");
+        S3Object object = s3.getObject(new GetObjectRequest(bucketName, keyName));
+        System.out.println("Content-Type: "  + object.getObjectMetadata().getContentType());
+
+
     }
 }
